Added/updated student dashboard controllers
