+++ app/controllers/student/regional_pitch_events_finder_controller.rb
@@ -1,0 +1,21 @@
+module Student
+  class RegionalPitchEventsFinderController < StudentController
+    include RequireParentalConsentSigned
+    include RequireLocationIsSet
+    def show
+      @regional_events = available_regional_events
+    end
+
+    private
+
+    def available_regional_events
+      if SeasonToggles.select_regional_pitch_event?
+        RegionalPitchEvent.available_to(
+          current_team.submission
+        )
+      else
+        RegionalPitchEvent.none
+      end
+    end
+  end
+end
