+++ app/controllers/student/team_creation_controller.rb
@@ -1,5 +1,5 @@
 module Student
-  class CurriculumController < StudentController
+  class TeamCreationController < StudentController
     include RequireParentalConsentSigned
     include RequireLocationIsSet
   end
