+++ app/controllers/student/team_submission_overview_controller.rb
@@ -1,5 +1,5 @@
 module Student
-  class CurriculumController < StudentController
+  class TeamSubmissionOverviewController < StudentController
     include RequireParentalConsentSigned
     include RequireLocationIsSet
   end
